acetwo-dashboard/backend/requirements.txt
fastapi
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9
python-dotenv
uvicorn[standard]>=0.20.0